updated navigation

+++ Localisation/src/ca/mcgill/ecse211/odometerlab/Coordinate.java
@@ -1,10 +1,10 @@
 package ca.mcgill.ecse211.odometerlab;
 
 public class Coordinate {
-	public int x;
-	public int y;
+	public double x;
+	public double y;
 	
-	public Coordinate(int x, int y) {
+	public Coordinate(double x, double y) {
 		this.x = x;
 		this.y = y;
 	}
